Update README.md
The computer needs to be running Python 2.7.
Operating system has to support Python -- preferably *nix. (select has acted funky for me in Windows before)
The game is best out of one game ("speed" Rock-Paper-Scissors).
Once a game is finished the winner and loser will be disconnected to make room for new users.